+++ src/main/java/dhbw/swingchat/helper/ThemedColors.java
@@ -2,15 +2,22 @@ package dhbw.swingchat.helper;
 
 import java.awt.Color;
 
-public class ThemedColors 
+/**
+ * Constants of colors for dark and light mode.
+ */
+public class ThemedColors
 {
-    private ThemedColors() { }
 
-    public static final Color darkForeground = Color.WHITE;
-    public static final Color darkBackround = Color.DARK_GRAY;
-    public static final Color darkSecondary = Color.GRAY;         //used for buttons, input, etc.
+    private ThemedColors()
+    {
+        // collection of constants
+    }
+
+    public static final Color darkForeground  = Color.WHITE;
+    public static final Color darkBackround   = Color.DARK_GRAY;
+    public static final Color darkSecondary   = Color.GRAY;      //used for buttons, input, etc.
 
     public static final Color lightForeground = Color.BLACK;
-    public static final Color lightBackround = Color.LIGHT_GRAY;
-    public static final Color lightSecondary = Color.WHITE;
+    public static final Color lightBackround  = Color.LIGHT_GRAY;
+    public static final Color lightSecondary  = Color.WHITE;
 }
